requirements.txt
asgiref==3.5.0
branca==0.5.0
certifi==2021.10.8
charset-normalizer==2.0.12
Django==4.0.4
django-admin-geomap==1.1
django-cors-headers==3.11.0
django-filter==21.1
django-leaflet==0.28.2
django-location-field==2.1.0
djangorestframework==3.13.1
djangorestframework-gis==1.0
djangorestframework-jsonapi==5.0.0
folium==0.12.1.post1
GDAL @ file:///../TerraBashkiria/GDAL-3.4.3-cp310-cp310-win_amd64.whl
idna==3.3
inflection==0.5.1
Jinja2==3.1.2
MarkupSafe==2.1.1
numpy==1.22.3
osgeo4w-installer==1.4.0
polyline==1.4.0
psycopg2==2.9.3
psycopg2-binary==2.9.3
pytz==2022.1
requests==2.27.1
six==1.16.0
sqlparse==0.4.2
tzdata==2022.1
urllib3==1.26.9
django-admin-interface==0.24.2